fixes DriveDistance logic & registers data
The isFinished() logic of DriveDistance was fixed to work while driving
backwards. If one wants to drive backwards, they should use a positive
distance and a negative speed. Data values for DriveDistance like
actualDistance, expectedDistance and stoppingDistance were registered
with TestingDashboard in the Drive Subsystem. A DriveDistanceSim command
was added to aid with simulations.